Split the test into blackbox and whitebox
